TECH-24 TECH-25 TECH-26 TECH-27 Remove dep on d2
