+++ examples/create-react-app/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@dhis2/d2-ui-app": "^1.0.0",
     "@dhis2/d2-ui-core": "^1.0.0",
     "@dhis2/d2-ui-expression-manager": "^1.0.0",
     "@dhis2/d2-ui-favorites-dialog": "^1.0.0",
