+++ examples/create-react-app/src/components/favorites-dialog.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import RaisedButton from 'material-ui/RaisedButton';
 import Snackbar from 'material-ui/Snackbar';
 
@@ -46,37 +44,35 @@ export default class FavoritesDialogExample extends React.Component {
     };
 
     render = () => (
-        <MuiThemeProvider muiTheme={getMuiTheme()}>
-            <div>
-                <div style={{ padding: 16 }}>
-                    <RaisedButton
-                        onClick={this.toggleDialog('chart')}
-                        label="Open charts favorites"
-                    />
-                </div>
-                <div style={{ padding: 16 }}>
-                    <RaisedButton
-                        onClick={this.toggleDialog('pivot')}
-                        label="Open pivot table favorites"
-                    />
-                </div>
-                <div style={{ padding: 16 }}>
-                    <RaisedButton onClick={this.toggleDialog('map')} label="Open maps favorites" />
-                </div>
-                <FavoritesDialog
-                    open={this.state.favoritesDialog.open}
-                    type={this.state.favoritesDialog.type}
-                    onRequestClose={this.toggleDialog(this.state.type)}
-                    onFavoriteSelect={this.onFavoriteSelect}
-                    d2={this.props.d2}
-                />
-                <Snackbar
-                    open={this.state.snackbar.open}
-                    message={this.state.snackbar.message}
-                    autoHideDuration={4000}
-                    onRequestClose={this.onSnackbarClose}
+        <div>
+            <div style={{ padding: 16 }}>
+                <RaisedButton
+                    onClick={this.toggleDialog('chart')}
+                    label="Open charts favorites"
                 />
             </div>
-        </MuiThemeProvider>
+            <div style={{ padding: 16 }}>
+                <RaisedButton
+                    onClick={this.toggleDialog('pivot')}
+                    label="Open pivot table favorites"
+                />
+            </div>
+            <div style={{ padding: 16 }}>
+                <RaisedButton onClick={this.toggleDialog('map')} label="Open maps favorites" />
+            </div>
+            <FavoritesDialog
+                open={this.state.favoritesDialog.open}
+                type={this.state.favoritesDialog.type}
+                onRequestClose={this.toggleDialog(this.state.type)}
+                onFavoriteSelect={this.onFavoriteSelect}
+                d2={this.props.d2}
+            />
+            <Snackbar
+                open={this.state.snackbar.open}
+                message={this.state.snackbar.message}
+                autoHideDuration={4000}
+                onRequestClose={this.onSnackbarClose}
+            />
+        </div>
     );
 }
