+++ examples/create-react-app/src/components/favorites-menu.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import Snackbar from 'material-ui/Snackbar';
 
 import FavoritesMenu from '@dhis2/d2-ui-favorites-menu';
@@ -41,29 +39,27 @@ export default class FavoritesMenuExample extends React.Component {
     };
 
     render = () => (
-        <MuiThemeProvider muiTheme={getMuiTheme()}>
-            <div>
-                <FavoritesMenu
-                    d2={this.props.d2}
-                    favoriteType="chart"
-                    onNew={this.onFavoritesMenuAction('New')}
-                    onOpen={this.onFavoritesMenuAction('Open')}
-                    onSave={this.onFavoritesMenuAction('Save')}
-                    onSaveAs={this.onFavoritesMenuAction('Save as')}
-                    onRename={this.onFavoritesMenuAction('Rename')}
-                    onTranslate={this.onFavoritesMenuAction('Translate')}
-                    onShare={this.onFavoritesMenuAction('Share')}
-                    onWriteInterpretation={this.onFavoritesMenuAction('Write interpretation')}
-                    onDelete={this.onFavoritesMenuAction('Delete')}
-                    onError={this.onFavoritesMenuError}
-                />
-                <Snackbar
-                    open={this.state.snackbar.open}
-                    message={this.state.snackbar.message}
-                    autoHideDuration={4000}
-                    onRequestClose={this.onSnackbarClose}
-                />
-            </div>
-        </MuiThemeProvider>
+        <div>
+            <FavoritesMenu
+                d2={this.props.d2}
+                favoriteType="chart"
+                onNew={this.onFavoritesMenuAction('New')}
+                onOpen={this.onFavoritesMenuAction('Open')}
+                onSave={this.onFavoritesMenuAction('Save')}
+                onSaveAs={this.onFavoritesMenuAction('Save as')}
+                onRename={this.onFavoritesMenuAction('Rename')}
+                onTranslate={this.onFavoritesMenuAction('Translate')}
+                onShare={this.onFavoritesMenuAction('Share')}
+                onWriteInterpretation={this.onFavoritesMenuAction('Write interpretation')}
+                onDelete={this.onFavoritesMenuAction('Delete')}
+                onError={this.onFavoritesMenuError}
+            />
+            <Snackbar
+                open={this.state.snackbar.open}
+                message={this.state.snackbar.message}
+                autoHideDuration={4000}
+                onRequestClose={this.onSnackbarClose}
+            />
+        </div>
     );
 }
