+++ examples/create-react-app/src/components/list-select.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 
 import {
     ListSelect,
@@ -19,50 +17,47 @@ const style = {
 
 export default function Buttons() {
     return (
-        <MuiThemeProvider muiTheme={getMuiTheme()}>
-            <div style={style}>
-                <div>
-                    <h3>List Select</h3>
-                    <ListSelect
-                        source={[
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                        ]}
-                        onItemDoubleClick={() => {}}
-                    />
-                </div>
-
-                <div>
-                    <h3>List Select Async</h3>
-                    <ListSelectAsync
-                        onItemDoubleClick={() => {}}
-                    />
-                </div>
-
-                <div>
-                    <h3>List Select With Local Search</h3>
-                    <ListSelectWithLocalSearch
-                        source={[
-                            {label: 'foo', value: 'bar'},
-                            {label: 'fuz', value: 'bar'},
-                            {label: 'faz', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                            {label: 'foo', value: 'bar'},
-                        ]}
-                        onItemDoubleClick={() => {}}
-                    />
-                </div>
-
+        <div style={style}>
+            <div>
+                <h3>List Select</h3>
+                <ListSelect
+                    source={[
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                    ]}
+                    onItemDoubleClick={() => {}}
+                />
             </div>
-        </MuiThemeProvider>
+
+            <div>
+                <h3>List Select Async</h3>
+                <ListSelectAsync
+                    onItemDoubleClick={() => {}}
+                />
+            </div>
+
+            <div>
+                <h3>List Select With Local Search</h3>
+                <ListSelectWithLocalSearch
+                    source={[
+                        {label: 'foo', value: 'bar'},
+                        {label: 'fuz', value: 'bar'},
+                        {label: 'faz', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                        {label: 'foo', value: 'bar'},
+                    ]}
+                    onItemDoubleClick={() => {}}
+                />
+            </div>
+        </div>
     );
 }
