+++ examples/create-react-app/src/components/tabs.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import { Tabs, Tab } from '@dhis2/d2-ui-core';
 
 const style = {
@@ -9,20 +7,18 @@ const style = {
 
 export default function TabsExample (props) {
     return (
-        <MuiThemeProvider muiTheme={getMuiTheme()}>
-            <div style={style}>
-                <Tabs>
-                    <Tab label='First'>
-                        First tab
-                    </Tab>
-                    <Tab label='Second'>
-                        Second tab
-                    </Tab>
-                    <Tab label='Third'>
-                        Third tab
-                    </Tab>
-                </Tabs>
-            </div>
-        </MuiThemeProvider>
+        <div style={style}>
+            <Tabs>
+                <Tab label='First'>
+                    First tab
+                </Tab>
+                <Tab label='Second'>
+                    Second tab
+                </Tab>
+                <Tab label='Third'>
+                    Third tab
+                </Tab>
+            </Tabs>
+        </div>
     );
 }
