+++ lerna.json
@@ -2,6 +2,7 @@
   "lerna": "2.9.0",
   "version": "independent",
   "npmClient": "yarn",
+  "npmClientArgs": ["--ignore-engines"],
   "useWorkspaces": true,
   "command": {
      "publish": {
