+++ packages/app/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@dhis2/d2-ui-core",
-  "version": "1.0.4",
-  "description": "Core components for DHIS2 in d2-ui",
+  "name": "@dhis2/d2-ui-app",
+  "version": "1.0.0",
+  "description": "D2-UI App wrappers",
   "main": "./build/cjs/index.js",
   "module": "./build/es/index.js",
   "files": [
@@ -14,22 +14,23 @@
     "test-ci": "jest --config=../../jest.config.js packages/core",
     "lint": "eslint src/",
     "prebuild": "rimraf ./build/*",
-    "build": "npm run build:commonjs && npm run build:es && npm run build:sass",
+    "build": "npm run build:commonjs && npm run build:es",
     "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir build/cjs --ignore spec.js",
     "build:es": "cross-env BABEL_ENV=es babel src --out-dir build/es --ignore spec.js",
     "build:umd": "cross-env && webpack -p",
-    "build:sass": "node-sass ./scss --output ./build/css",
-    "watch": "npm run build:sass && npm run build:es --  --watch"
+    "watch": "npm run build:es --  --watch"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/dhis2/d2-ui.git"
   },
   "dependencies": {
+    "@dhis2/d2-ui-core": "^1.0.4",
     "lodash": "^4.17.10",
     "material-ui": "^0.20.0"
   },
   "peerDependencies": {
+    "d2": "^30.0.0",
     "react": "^16.0.0",
     "react-dom": "^16.0.0"
   },
